readonly counter update

+++ src/utils/ReadonlyCounter.cpp
@@ -1,7 +1,9 @@
+#include <cstdint>
+
 #include <utils/ReadonlyCounter.h>
 
 namespace arch::utils {
 
-template class ReadonlyCounter<size_t>;
+template class ReadonlyCounter<uint32_t>;
 
 }
